Major Push
I figured I need to push something, still in the process of getting things to work so I was trying not to push it but no uploads for a long time looks very bad. Not all of my work or a final push; currently trying to get some other things in the program working before I push more. Should be pushing again before the end of today (12/7/22).